api-metadata/validation/python/tests/requirements.txt
pytest>=7.0
requests>=2.28.0
pyyaml>=6.0
